fix: Resolve all Dart type errors and warnings
- Fix return type mismatch in extensions.dart (num to double)
- Fix nullable String handling in shopping_list_provider.dart
- Fix parameter naming in manual_food_entry_screen.dart
- Fix multiple nullable access errors in meal_plan_detail_screen.dart
- Fix num to int assignment in meal_plans_screen.dart
- Fix undefined getter 'name' to 'ingredientName' in pantry_screen.dart
- Remove unnecessary null comparisons and assertions
- Make _favoriteRecipes field final in recipe_provider.dart
- Remove unused _searchQuery field from pantry_screen.dart
- Remove unnecessary toList() in spread operator usage
- Add mounted check for async context usage in settings_screen.dart

+++ lib/presentation/screens/meal_plans/meal_plans_screen.dart
@@ -390,7 +390,7 @@ class _MealPlansScreenState extends State<MealPlansScreen> {
   int _calculateTotalMeals(plan) {
     int total = 0;
     for (final day in plan.days) {
-      total += day.meals.length;
+      total += day.meals.length as int;
     }
     return total;
   }
